TravisCI Test
This commit includes a basic test that the site builds and that html-proofer
doesn't return any errors. It was based on the following documentation:

http://jekyllrb.com/docs/continuous-integration/travis-ci/

Ideally there would be something that checked that the datasets.yml file has the
expected elements. But this will at least make sure there are no errors that
cause the site to silently fail to build. Now Travis will run the test on
commits and make an announcement in Slack.

Fixes #7