Data Conversion + Gatsby
This commit adds convert.py that reads the old datasets.yml and writes
it as a set of markdown files in src/datasets/ Some data massaging
needed to be done, especially to supply missing dates.

Also included in this commit is a barebone Gatsby application which is
replacing Jekyll so that it can work more like an web application
(sorting, filtering, etc).