netlify-cms
This commit adds an initial netlify-cms configuration to allow people to
submit datasets while keeping the site as static.

Note: we discussed adding email addresses for people in case dataset
users want to get in contact with the dataset creator(s). So convert.py
was updated to have creators be a list of objects that have a name and
email.